API Reference

MRICompress.block_svt — Method

block_svt(z, c, patch_sz, num_iterations, lambda::Float64, overlap_y::Int64)

Reconstructs an image using weighted overlapped patches where an SVT is applied to each patch. Patch size and overlap amount should be specified such that indexation errors are avoided. Returns a compressed image using blocked singular value thresholding.

MRICompress.kspace_to_img — Method

kspace_to_img(img_path)

Converts an image in k-space to pixel values via the IFFT.

MRICompress.split_bregman_anisotropic_tv_denoising — Method

split_bregman_anisotropic_tv_denoising(img, lambda, mu, tol, max_iterations)

Implementation based on The Split Bregman Method For L1 Regularized Problems Goldstein, Osher Returns a denoised image. Note: tol currently unused.

Examples

julia> img = kspace_to_img();
julia> split_bregman_anisotropic_tv_denoising(img, 1, 1, 0, 1);

MRICompress.svt — Method

svt(img, lambda, num_iterations)

Returns a compressed image using singular value thresholding.

MRICompress.wiener_deconv — Method

wiener_deconv(img, orig, sigma)

Returns a denoised image.