Changelog

All notable changes to this project will be documented in this file.

The format is based on Keep a Changelog, and this project adheres to Semantic Versioning.

Base.:- — Function

-(x)

Unary minus operator.

See also: abs, flipsign.

Examples

julia> -1
-1

julia> -(2)
-2

julia> -[1 2; 3 4]
2×2 Matrix{Int64}:
 -1  -2
 -3  -4

-(x, y)

Subtraction operator.

Examples

julia> 2 - 3
-1

julia> -(2, 4.5)
-2.5

Missing docstring.

Missing docstring for `### Changed

-`. Check Documenter's build log for details.

Missing docstring.

Missing docstring for `### Removed

-`. Check Documenter's build log for details.